+++ RELEASE-NOTES.md
@@ -1,6 +1,7 @@
 ## 3.0.M1
 
 * #1 Support "recipientOverride" setting
+* #3 Support "simple-java-mail" 7.4.0 (Jakarta namespace)
 * #4 Upgrade to "simple-java-mail" 6.7.6
 
 ## 2.0.M1